create user session ID ourselves
