+++ alphastats/gui/utils/ui_helper.py
@@ -1,3 +1,5 @@
+import uuid
+
 import streamlit as st
 import pandas as pd
 import base64
@@ -8,7 +10,7 @@ from alphastats import __version__
 
 
 def sidebar_info():
-    display_sidebar_html_table()
+    _display_sidebar_html_table()
     st.sidebar.markdown("\n\n")
     st.sidebar.markdown("AlphaPeptStats Version " + str(__version__))
     st.sidebar.info(
@@ -34,7 +36,7 @@ def sidebar_info():
     )
 
 
-def display_sidebar_html_table():
+def _display_sidebar_html_table():
     if "dataset" not in st.session_state:
         return
 
@@ -58,3 +60,24 @@ def img_to_bytes(img_path):
     # img_bytes = Path(img_path).read_bytes()
     # encoded = base64.b64encode(img_bytes).decode()
     return encoded_string.decode()
+
+
+def empty_session_state():
+    """
+    remove all variables to avoid conflicts
+    """
+    for key in st.session_state.keys():
+        del st.session_state[key]
+    st.empty()
+
+
+def init_session_state() -> None:
+    """Initialize the session state."""
+
+    st.session_state["user_session_id"] = str(uuid.uuid4())
+
+    if "gene_to_prot_id" not in st.session_state:
+        st.session_state["gene_to_prot_id"] = {}
+
+    if "organism" not in st.session_state:
+        st.session_state["organism"] = 9606  # human
